+++ src/Network/Network/BYOIP/CustomIpPrefix/NewAzureCustomIpPrefixCommand.cs
@@ -21,6 +21,7 @@ namespace Microsoft.Azure.Commands.Network
     using System.Collections;
     using System.Linq;
     using System.Management.Automation;
+    using System.Text.RegularExpressions;
     using MNM = Microsoft.Azure.Management.Network.Models;
 
     [Cmdlet(VerbsCommon.New, ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "CustomIpPrefix", SupportsShouldProcess = true), OutputType(typeof(PSCustomIpPrefix))]
@@ -99,9 +100,12 @@ namespace Microsoft.Azure.Commands.Network
 
         [Parameter(
             Mandatory = false,
-            HelpMessage = "Parent CustomIpPrefix for /64 IPv6 CustomIpPrefix",
+            HelpMessage = "Parent CustomIpPrefix for Child CustomIpPrefix",
             ValueFromPipelineByPropertyName = true)]
         public PSCustomIpPrefix CustomIpPrefixParent { get; set; }
+
+        [Parameter(Mandatory = false, HelpMessage = "Denotes that resource is being created as a Parent CustomIpPrefix")]
+        public SwitchParameter IsParent { get; set; }
 
         [Parameter(
             Mandatory = false,
@@ -152,6 +156,22 @@ namespace Microsoft.Azure.Commands.Network
                 ExpressRouteAdvertise = this.ExpressRouteAdvertise
             };
 
+            if (IsIPv4CIDR())
+            {
+                if (this.IsParent)
+                {
+                    psModel.PrefixType = "Parent";
+                }
+                else if (this.CustomIpPrefixParent != null)
+                {
+                    psModel.PrefixType = "Child";
+                }
+                else
+                {
+                    psModel.PrefixType = "Singular";
+                }
+            }
+
             var sdkModel = NetworkResourceManagerProfile.Mapper.Map<MNM.CustomIpPrefix>(psModel);
 
             sdkModel.Tags = TagsConversionHelper.CreateTagDictionary(this.Tag, validate: true);
@@ -166,6 +186,18 @@ namespace Microsoft.Azure.Commands.Network
             }
 
             return null;
+        }
+
+        private bool IsIPv4CIDR()
+        {
+            if (this.Cidr != null)
+            {
+                return Regex.IsMatch(this.Cidr, @"^\d+\.\d+\.\d+\.\d+/\d+$");
+            }
+            else
+            {
+                return false;
+            }
         }
     }
 }